Added e2e test for keyboard context menu
